requirements.txt
click==8.1.3
ipython==8.14.0
pandas==2.0.1
python-dotenv==1.0.0
scikit_learn==1.2.2
seqeval==1.2.2
setuptools==67.6.1
docopt-ng==0.8.1
spacy==3.5.0
nltk==3.8.1
spacy-transformers==1.2.4
pytest==7.3.2
annoy==1.17.3
sentence_transformers==2.2.2
gensim==4.3.1
torch==2.0.1
dash==2.11.0
dash_bootstrap_components==1.4.1
dash_cytoscape==0.3.0
matplotlib==3.5.0
networkx==2.6.3
plotly==5.15.0
gunicorn>=19.9.0